Dramatically decreased stack size needed for vagga_build
Decrease from ~34k -> ~8k (should partiall fix #180, and maybe #161)
