+++ src/builder/mod.rs
@@ -24,6 +24,11 @@ mod commands {
     pub mod npm;
     pub mod vcs;
     pub mod download;
+    pub mod subcontainer;
+    pub mod copy;
+    pub mod text;
+    pub mod dirs;
+    pub mod packaging;
 }
 mod capsule;
 mod packages;
